+++ demo/__tests__/style.test.tsx
@@ -2,6 +2,10 @@ import { render } from '@testing-library/react';
 
 import App from '../App';
 
+function getDocumentHTML() {
+  return document.documentElement.outerHTML.replace(/\r\n|\n|\r/g, '\n');
+}
+
 describe('Style', () => {
   it('should render CSS correctly in JSDOM', () => {
     render(<App />);
@@ -10,20 +14,26 @@ describe('Style', () => {
     // Global CSS
     // TODO: Global CSS is saved into the `.cache` folder, so we can't assert it directly within the JSDOM
     // Imported CSS
-    expect(document.documentElement.outerHTML).toContain(
-      '<link rel="stylesheet" href="/demo/App.css">',
+    expect(getDocumentHTML()).toContain(
+      `.App {
+  text-align: center;
+}`,
     );
-    expect(document.documentElement.outerHTML).toContain(
-      '<link rel="stylesheet" href="/demo/assets/css/App.css">',
+    expect(getDocumentHTML()).toContain(
+      `.logo2 {
+  max-width: 300px;
+  background: yellow;
+}`,
     );
 
     // styled-components
-    expect(document.documentElement.outerHTML).toContain(
-      '<style data-styled="active" data-styled-version="5.3.5">.dgihId{color:red;}</style>',
+    expect(getDocumentHTML()).toContain(
+      '<style data-styled="active" data-styled-version="5.3.5">',
     );
+    expect(getDocumentHTML()).toContain('.dgihId{color:red;}');
 
     // emotion
-    expect(document.documentElement.outerHTML).toContain(
+    expect(getDocumentHTML()).toContain(
       '<style data-emotion="css" data-s="">.css-2m18qq{color:orange;}</style>',
     );
 
@@ -31,35 +41,25 @@ describe('Style', () => {
     // Global
     // TODO: not implemented yet
     // Import
-    expect(
-      document.documentElement.outerHTML.replace(/\r\n|\n|\r/g, ''),
-    ).toContain(
-      `<style type=\"text/css\">._cssModule_1gc0y_1 {  color: green;}</style>`,
+    expect(getDocumentHTML()).toContain(
+      `._cssModule_1a0qn_1 {
+  color: green;
+}`,
     );
 
     // Sass
     // Global
     // TODO: Global SCSS is saved into the `.cache` folder, not in the JSDOM, need to find a way to test it
     // Import
-    expect(document.documentElement.outerHTML)
-      .toContain(`header .imported-sass {
+    expect(getDocumentHTML()).toContain(`header .imported-sass {
   text-transform: uppercase;
 }`);
-    expect(document.documentElement.outerHTML)
-      .toContain(`header .imported-sass {
+    expect(getDocumentHTML()).toContain(`header .imported-sass {
   color: pink;
-}</style>`);
+}`);
     // import ~
     // TODO: Not implemented yet
     // loadPaths
     // TODO: Not implemented yet
-
-    // Can see images
-    expect(document.documentElement.outerHTML).toContain(
-      `<img src="/logo.svg" class="App-logo" alt="logo">`,
-    );
-    expect(document.documentElement.outerHTML).toContain(
-      `<img src="/demo/assets/images/logo.svg" class="logo2" alt="logo2">`,
-    );
   });
 });