+++ examples/svelte/jest.config.js
@@ -1,0 +1,20 @@
+/** @type {import('@jest/types').Config.InitialOptions} */
+
+export default {
+  setupFilesAfterEnv: ['<rootDir>/src/setupTests.ts'],
+  transform: {
+    '^.+\\.ts$': 'ts-jest',
+    '^.+\\.svelte$': [
+      'svelte-jester',
+      {
+        preprocess: true,
+      },
+    ],
+    // Read more at https://www.jest-preview.com/docs/getting-started/installation
+    '^.+\\.(css|scss|sass|less)$': 'jest-preview/transforms/css',
+    '^(?!.*\\.(js|jsx|mjs|cjs|ts|tsx|css|json)$)':
+      'jest-preview/transforms/file',
+  },
+  moduleFileExtensions: ['js', 'ts', 'svelte'],
+  testEnvironment: 'jsdom',
+};